move special case of DataTakingException in debug-mode into main().
Imho, the debug option is a source of possible confusion and
certainly adds complexity. It should (if possible) be handled only
in the main script, i.e. the project should not be full of
special cases, which are hard to keep in mind.

I must even say, that we should get rid of this option
once, if possible. It is too extraordinary, to have a debug mode
which is not, e.g. slower and producing more log-output,
but which *behaves* differently.

It would be much more usual, I guess, to have a mock-up Dim-server,
where a mock-up DIM_CONTROL is always in 'Running',
and other servers like WEATHER and FEEDBACK are producing
test service-data, which can be used for testing...
But unfortunately, we don't have that.