add new IntervalCheck for new condition
This check should only fire, if there is an endless loop in main.js again. These were in the past solved by experts. But maybe we should run them through the shifter first, and let them try to fix it with a simple stop-restart. 

This check can potentially lead to false positives during startup ... but we have to see this.